+++ src/StackNucleus.DDD.Domain/Images/ImageSize.cs
@@ -1,0 +1,20 @@
+﻿namespace StackNucleus.DDD.Domain.Images;
+
+/// <summary>
+/// Represents a size of an image
+/// </summary>
+public enum ImageSize
+{
+    /// <summary>
+    /// Small size
+    /// </summary>
+    Small = 0,
+    /// <summary>
+    /// Medium size
+    /// </summary>
+    Medium = 1,
+    /// <summary>
+    /// Large size
+    /// </summary>
+    Large = 2
+}