+++ src/StackNucleus.DDD.Domain/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 This project follows [Semantic Versioning (SemVer)](https://semver.org/).
 
+## [2.1.0] - 2025-09-15
+
+### Changed
+
+- Image uploader and compressor location.
+- Improved API contract for the image uploader and compressor.
+
+### Added
+
+- Contract to upload multiple file sizes per one request to the image uploader.
+
 ## [2.0.0] - 2025-09-11
 
 ### Changed